+++ src/chip8/registers.rs
@@ -2,7 +2,7 @@ pub struct Registers {
     pub v: Vec<u8>,
     pub i: u16,
     pub pc: u16,
-    pub sp: u8,
+    pub sp: usize,
     pub stack: Vec<u16>,
     /// delay timer
     pub dt: u8,
